form-handling/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
attrs==24.2.0
autopep8==2.3.1
backports.tarfile==1.2.0
bleach==6.1.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
cssselect==1.2.0
decorator==5.1.1
dicttoxml==1.7.16
dill==0.3.8
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.5
fastapi-mail==1.4.1
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
geoip2==4.8.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
ip2geotools==0.1.6
IP2Location==8.10.2
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.4
keyring==25.3.0
lazy-object-proxy==1.10.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
maxminddb==2.6.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.4.0
multidict==6.0.5
nh3==0.2.18
outcome==1.3.0.post0
packaging==24.1
pip-review==1.3.0
pkginfo==1.10.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
pylint==3.2.6
pyparsing==3.1.2
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
ratelim==0.1.6
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
SecretStorage==3.3.3
selenium==4.23.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
sqlmodel==0.0.21
starlette==0.37.2
tomlkit==0.13.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
twine==5.1.1
typed-ast==1.5.5
typer==0.12.3
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.0
